+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- When creating the MC OIDC provider on the WC account, use the S3 URL when in China regions.
+
 ## [0.31.0] - 2025-03-20
 
 ### Changed